requirements.txt
tensorflow_probability==0.3.0
tensorflow-gpu==1.10.0
dm-sonnet==1.23
matplotlib==2.2.2
numpy==1.14.5
Pillow==5.1.0
scipy==1.1.0
SimpleITK==1.1.0
tensorboard==1.10.0
tqdm==4.23.4
seaborn==0.9.0
pytest==3.8.0